requirements.txt
Flask==2.3.2
itsdangerous==2.1.2  
Jinja2==3.1.2       
MarkupSafe==2.1.3   
Werkzeug==2.3.6    
pdf2image==1.16.3
Pillow==9.3.0       
transformers==4.31.0
huggingface-hub==0.16.4  
tokenizers==0.13.3       
tensorflow==2.13.0
torch==2.0.1
tqdm==4.66.1             
numpy     
requests==2.31.0 
protobuf==4.24.0  
packaging==23.1 
